+++ app/Validators/discords/StoreValidator.ts
@@ -1,7 +1,7 @@
 import {rules, schema} from '@ioc:Adonis/Core/Validator'
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 
-export default class UpdateValidator {
+export default class StoreValidator {
   constructor (protected ctx: HttpContextContract) {
   }
 
@@ -25,17 +25,10 @@ export default class UpdateValidator {
 	 *    ```
 	 */
   public schema = schema.create({
-    username: schema.string.optional({trim: true}),
-    email: schema.string.optional({ trim: true }, [rules.email(), rules.unique({ table: 'users', column: 'email' })]),
-    password: schema.string.optional({ trim: true }, [rules.confirmed()]),
-    credit: schema.number.optional(),
-
-    banned: schema.boolean.optional(),
-    linked: schema.boolean.optional(),
-
-    roles: schema.array.optional().members(schema.string({trim: true}, [rules.exists({column: 'id', table: 'roles'})])),
-    permissions: schema.array.optional().members(schema.string()),
-    discord: schema.array.optional().members(schema.string()),
+    discord_id: schema.string({trim: true}, [rules.unique({table: 'discords', column: 'discord_id'})]),
+    level: schema.number(),
+    exp: schema.number(),
+    username: schema.string({trim: true})
   })
 
 	/**
